firewire-motu: wait for notification when changing clock configuration for protocol v3
It costs expensive to change clock configuration for models of protocol
version 3. In current implementation, speculative strategy is used for
the operation; just waiting for 4 seconds. As long as I investigate,
when accepting and changing clock status actually, the device sends
notification with mask.

This commit uses wise way to wait for the notification at changing
sampling clock rate with timeout of 4 seconds.